Move permutation info to separate file
